requirements.txt
aiogram
sqlalchemy
aiosqlite
apscheduler
python-dotenv
groq
pytz
